# openapi.yaml
openapi: 3.0.0
info:
  title: ShopBuilder API
  version: 2.0.0
  description: >-

    ## ShopBuilder — Enterprise Multi-Tenant E-Commerce Platform


    A scalable multi-tenant backend supporting multiple merchants, stores, and customers within isolated tenant
    namespaces.


    ### Role-Based Access Control (RBAC)


    | Role | Description |

    |------|-------------|

    | `SUPER_ADMIN` | Full platform access — manages tenants, users, everything |

    | `PLATFORM_ADMIN` | Manages merchants and platform configuration |

    | `MERCHANT_OWNER` | Manages own stores, products, and orders |

    | `STORE_MANAGER` | Manages inventory and orders for assigned store |

    | `CUSTOMER` | Places and views own orders |


    ### Authentication

    Include a Bearer token on all protected endpoints:

    ```

    Authorization: Bearer <access_token>

    ```


    Access tokens expire in **15 minutes**. Use `POST /auth/refresh` with your refresh token to get a new one.


    ### Multi-Tenancy

    All business data is scoped to a `tenantId`. It is embedded in the JWT and automatically applied.

    Admins can override `tenantId` via query parameter.
          
  contact:
    name: ShopBuilder API Support
    email: api@shopbuilder.io
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login
...
security:
  - bearerAuth: []
paths:
...
  /inventory/{variantId}/status:
    get:
      tags:
        - Inventory
      summary: Get inventory status for a variant
      description: Returns total stock, reserved stock, and computed available stock.
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Inventory status
          content:
            application/json:
              example:
                variantId: uuid
                sku: PROD-S-Black-Cotton
                stock: 100
                reservedStock: 15
                availableStock: 85
        '404':
          description: Variant not found
  /inventory/{variantId}/increase:
    post:
      tags:
        - Inventory
      summary: Increase stock (goods received)
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  example: 50
      responses:
        '200':
          description: Stock increased
        '400':
          description: Invalid quantity
        '404':
          description: Variant not found
  /inventory/{variantId}/decrease:
    post:
      tags:
        - Inventory
      summary: Decrease stock (shrinkage, damage, write-off)
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  example: 3
      responses:
        '200':
          description: Stock decreased
        '400':
          description: Insufficient stock
  /inventory/{variantId}/reserve:
    post:
      tags:
        - Inventory
      summary: Reserve stock (hold for pending order)
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  example: 5
      responses:
        '200':
          description: Stock reserved
        '400':
          description: Insufficient available stock
  /inventory/{variantId}/release:
    post:
      tags:
        - Inventory
      summary: Release reserved stock (order cancelled or expired)
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  example: 5
      responses:
        '200':
          description: Reserved stock released
        '400':
          description: Cannot release more than reserved
  /inventory/transfer:
    post:
      tags:
        - Inventory
      summary: Transfer stock between two variants (warehouse transfer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fromVariantId
                - toVariantId
                - quantity
              properties:
                fromVariantId:
                  type: string
                  format: uuid
                toVariantId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
                  example: 10
      responses:
        '200':
          description: Transfer completed (atomic transaction)
          content:
            application/json:
              example:
                message: 10 units transferred
                transferred: 10
                from:
                  id: uuid
                  sku: A
                  stock: 40
                to:
                  id: uuid
                  sku: B
                  stock: 60
        '400':
          description: Insufficient source stock or same variant
  /merchants:
    get:
      tags:
        - Merchants
      summary: List merchants (SUPER_ADMIN, PLATFORM_ADMIN, MERCHANT_OWNER)
      parameters:
        - in: query
          name: tenantId
          schema:
            type: string
          description: Filter by tenant (admin only)
        - in: query
          name: status
          schema:
            type: string
            enum:
              - ACTIVE
              - SUSPENDED
              - PENDING
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated merchant list
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags:
        - Merchants
      summary: Create a merchant (SUPER_ADMIN, PLATFORM_ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - tenantId
              properties:
                name:
                  type: string
                  example: Acme Retail Ltd
                email:
                  type: string
                  format: email
                  example: contact@acme.com
                phone:
                  type: string
                  example: +1-555-0100
                address:
                  type: string
                  example: 123 Commerce St, NY 10001
                tenantId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Merchant created
        '409':
          description: Email already in use
  /merchants/{id}:
    get:
      tags:
        - Merchants
      summary: Get merchant by ID (with stores)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Merchant detail
        '404':
          description: Merchant not found
    patch:
      tags:
        - Merchants
      summary: Update merchant details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                address:
                  type: string
      responses:
        '200':
          description: Updated merchant
        '404':
          description: Merchant not found
    delete:
      tags:
        - Merchants
      summary: Delete a merchant (SUPER_ADMIN only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Merchant deleted
        '404':
          description: Merchant not found
  /merchants/{id}/suspend:
    patch:
      tags:
        - Merchants
      summary: Suspend a merchant account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Merchant suspended
  /merchants/{id}/activate:
    patch:
      tags:
        - Merchants
      summary: Reactivate a suspended merchant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Merchant activated
...